Started to implement Authentication
Added the auth command

Turns out we can get the message's server by doing "msg.channel.server"
so, I've added that to the commands that uses the server. Might change
functions to just take the "msg" object and get the server via that
(improve simplicity)

Config now has a "auth_ip" which is shown to users when they just the
auth command without arguments.