Team command update
Updated the team command to work with DB.

Added a delete and list command (both admin only).

Create command created a team in the DB, sets the user as the owner
(also updates their teamName property).

Fixed "undefined" showing as team name when running whois.

Modified the schemas so the models store Strings (representing
discordIds) instead of the user model (may be faster?)